Removing all gather workaround
