+++ test/python/golden/test_ttir_ops.py
@@ -1981,95 +1981,6 @@ def test_mesh_shard_devices(
 @pytest.mark.parametrize(
     "test_shape",
     [
-        (1, 32, 32, 32),
-        (1, 32, 32, 1),
-        (32, 32, 1, 1),
-        (1, 32, 32),
-        (32, 32),
-        (32, 40),
-        (40, 32),
-        pytest.param((1, 1, 32, 32, 32), marks=pytest.mark.xfail(reason="run error")),
-        pytest.param(
-            (1, 1, 1, 1, 1, 1, 32, 32, 32), marks=pytest.mark.xfail(reason="run error")
-        ),
-    ],
-    ids=shape_str,
-)
-@pytest.mark.parametrize(
-    "mesh_shape", [(2, 4), (1, 8), (1, 2), (1, 32), (8, 4)], ids=shape_str
-)
-@pytest.mark.parametrize("all_gather_dim", range(4))
-@pytest.mark.parametrize("cluster_axis", [0, 1])
-@pytest.mark.parametrize("dtype", [torch.bfloat16, torch.float32], ids=["bf16", "f32"])
-def test_all_gather(
-    test_shape: Shape,
-    mesh_shape: Tuple[int, int],
-    all_gather_dim: int,
-    cluster_axis: int,
-    dtype: torch.dtype,
-    request,
-    device,
-):
-    if all_gather_dim >= len(test_shape):
-        pytest.skip("all_gather_dim is out of range")
-    if mesh_shape[cluster_axis] == 1:
-        pytest.skip("all_gather across 1 device is meaningless")
-
-    rank_in = len(test_shape)
-    rank_mesh = len(mesh_shape)
-
-    if rank_mesh > rank_in:
-        raise ValueError(
-            f"Mesh shape {mesh_shape} has {rank_mesh} dimensions, but test shape "
-            f"{test_shape} only has {rank_in} dimensions. Cannot shard more "
-            f"dimensions than exist in the tensor."
-        )
-
-    # Take the last `rank_mesh` dims as sharded dims
-    shard_dims = list(range(rank_in - rank_mesh, rank_in))
-    shard_shape = make_shard_shape(rank_in, shard_dims, mesh_shape)
-
-    full_input_shape = list(test_shape)
-    for d, factor in zip(shard_dims, mesh_shape):
-        full_input_shape[d] *= factor
-
-    def module(builder: TTIRBuilder):
-        @builder.func([full_input_shape], [dtype])
-        def all_gather(in0: Operand, builder: TTIRBuilder):
-            in_shard = builder.mesh_shard(
-                in0,
-                shard_direction=MeshShardDirection.FullToShard.value,
-                shard_type=MeshShardType.Devices.value,
-                shard_shape=shard_shape,
-                shard_dims=shard_dims,
-            )
-
-            all_gather0 = builder.all_gather(
-                in_shard,
-                all_gather_dim=all_gather_dim,
-                cluster_axis=cluster_axis,
-            )
-
-            return builder.mesh_shard(
-                all_gather0,
-                shard_direction=MeshShardDirection.ShardToFull.value,
-                shard_type=MeshShardType.Devices.value,
-                shard_shape=shard_shape,
-                shard_dims=shard_dims,
-            )
-
-    compile_and_execute_ttir(
-        module,
-        mesh_name="mesh",
-        device=device,
-        mesh_dict=OrderedDict([("x", mesh_shape[0]), ("y", mesh_shape[1])]),
-        **get_request_kwargs(request),
-    )
-
-
-@pytest.mark.parametrize(
-    "test_shape",
-    [
         (1, 1, 256, 256),
         (256, 256, 1, 1),
         (1, 64, 64),
