+++ test/ttmlir/Dialect/TTNN/ccl/all_gather/all_gather_positive.mlir
@@ -25,39 +25,3 @@ module attributes {} {
     return %0 : tensor<1x1x32x128xbf16>
   }
 }
-
-// -----
-
-// Verify that the we reshape the input of the all_gather op, on 2D tensors
-module attributes {} {
-  // CHECK-LABEL: all_gather_reshape_2D
-  func.func @all_gather_reshape_2D(%arg0: tensor<32x128xbf16>) -> tensor<128x128xbf16> {
-    %0 = "ttir.all_gather"(%arg0) <{all_gather_dim = 0 : si32, cluster_axis = 1 : ui32}> : (tensor<32x128xbf16>) -> tensor<128x128xbf16>
-    // CHECK: "ttnn.reshape"
-    // CHECK-SAME: shape = [1 : i32, 1 : i32, 32 : i32, 128 : i32]
-    // CHECK-SAME: tensor<32x128xbf16
-    // CHECK: "ttnn.all_gather"
-    // CHECK: "ttnn.reshape"
-    // CHECK-SAME: {shape = [128 : i32, 128 : i32]}
-    // CHECK-SAME: tensor<128x128xbf16
-    return %0 : tensor<128x128xbf16>
-  }
-}
-
-// -----
-
-// Verify that the we reshape the input of the all_gather op, on 1D tensors
-module attributes {} {
-  // CHECK-LABEL: all_gather_reshape_1D
-  func.func @all_gather_reshape_1D(%arg0: tensor<32xbf16>) -> tensor<128xbf16> {
-    %0 = "ttir.all_gather"(%arg0) <{all_gather_dim = 0 : si32, cluster_axis = 1 : ui32}> : (tensor<32xbf16>) -> tensor<128xbf16>
-    // CHECK: "ttnn.reshape"
-    // CHECK-SAME: shape = [1 : i32, 1 : i32, 1 : i32, 32 : i32]
-    // CHECK-SAME: tensor<32xbf16
-    // CHECK: "ttnn.all_gather"
-    // CHECK: "ttnn.reshape"
-    // CHECK-SAME: {shape = [128 : i32]}
-    // CHECK-SAME: tensor<128xbf16
-    return %0 : tensor<128xbf16>
-  }
-}
